Refactor transcription stack into modular package
